(1) SPIA2D.h: force type change of adChn when use adChn as a subscript of array that requires an int type. (2) Makefile: add OBJECTS_DOS and OBJECT_DOS_ALL to enable the cs-make clean works. Convert UNIX string into DOS string (/ <-> \) use DOS's del command to remove files rather than UNIX rm

+++ SPIA2D.h
@@ -121,7 +121,7 @@ void Intr_SSTRB(void)
      a2dtemp=hi<<8;
      a2dtemp=a2dtemp+lo;
      a2dtemp=a2dtemp>>3;
-     a2dvalue[nSamples][adChn]=a2dtemp;
+     a2dvalue[nSamples][(unsigned int)adChn]=a2dtemp;
      if (adChn>=maxChnID)
         { 
           adChn=0;
